Mini cambios
Mini cambios para hacer similar la experencia

+++ Proyecto_G5_P3_Android/SistemaBienestarEstudiantil/app/src/main/java/espol/poo/sistemabienestarestudiantil/modelo/calculadora/CalculadoraPromedio.java
@@ -1,10 +1,12 @@
 package espol.poo.sistemabienestarestudiantil.modelo.calculadora;
 
+
 public class CalculadoraPromedio {
-    private double parcial1, parcial2, practico, mejoramiento;
+    // Las notas de entrada son enteras según tu requerimiento
+    private int parcial1, parcial2, practico, mejoramiento;
     private int porcentajeTeorico;
 
-    public CalculadoraPromedio(double p1, double p2, double prac, double mej, int porc) {
+    public CalculadoraPromedio(int p1, int p2, int prac, int mej, int porc) {
         this.parcial1 = p1;
         this.parcial2 = p2;
         this.practico = prac;
@@ -13,10 +15,10 @@ public class CalculadoraPromedio {
     }
 
     public double calcularNotaFinal() {
-        double notaP1 = parcial1;
-        double notaP2 = parcial2;
+        int notaP1 = parcial1;
+        int notaP2 = parcial2;
 
-        // Mejoramiento reemplaza la más baja solo si es mayor
+        // Lógica de mejoramiento: reemplaza la más baja solo si la mejora
         if (mejoramiento > parcial1 || mejoramiento > parcial2) {
             if (parcial1 <= parcial2) {
                 notaP1 = mejoramiento;
@@ -25,15 +27,18 @@ public class CalculadoraPromedio {
             }
         }
 
+        // Promedio teórico (usamos 2.0 para forzar que el resultado sea decimal)
         double promedioTeorico = (notaP1 + notaP2) / 2.0;
+
         double pesoTeorico = porcentajeTeorico / 100.0;
         double pesoPractico = 1.0 - pesoTeorico;
 
+        // Retorna el cálculo exacto con decimales
         return (promedioTeorico * pesoTeorico) + (practico * pesoPractico);
     }
 
     public String determinarEstado(double notaFinal) {
-        // CAMBIO: Ahora comparamos contra 60.0
+        // Regla de oro: 60.00 es el límite. 59.99 es reprobado.
         return (notaFinal >= 60.0) ? "¡Aprobaste! :D" : "Reprobaste :c";
     }
 }
